+++ packages/react/src/components/Tag/index.tsx
@@ -4,6 +4,7 @@ import classNames from 'classnames';
 interface TagProps {
   children: React.ReactNode;
   className?: string;
+  size?: 'default' | 'small';
 }
 
 export const TagLabel = ({ children, className, ...other }: TagProps) => (
@@ -13,8 +14,13 @@ export const TagLabel = ({ children, className, ...other }: TagProps) => (
 );
 TagLabel.displayName = 'TagLabel';
 
-const Tag = ({ children, className, ...other }: TagProps) => (
-  <div className={classNames('Tag', className)} {...other}>
+const Tag = ({ children, className, size = 'default', ...other }: TagProps) => (
+  <div
+    className={classNames('Tag', className, {
+      'Tag--small': size === 'small'
+    })}
+    {...other}
+  >
     {children}
   </div>
 );